suppress conversion warnings
-- tolower takes chars and returns chars, types are int because C only deals with ints
-- format parameter is size_t but there is no std parsing function that returns size_t. stoull is used instead as the widest alternative, and then the cast is applied to suppress the conversion warnings

+++ source/detail/serialization/xlsx_consumer.cpp
@@ -246,7 +246,7 @@ cell xlsx_consumer::read_cell()
 
     if (parser().attribute_present("s"))
     {
-        cell.format(target_.format(std::stoull(parser().attribute("s"))));
+        cell.format(target_.format(static_cast<std::size_t>(std::stoull(parser().attribute("s")))));
     }
 
     auto has_value = false;
@@ -673,7 +673,7 @@ void xlsx_consumer::read_worksheet_sheetdata()
 
             if (parser().attribute_present("s"))
             {
-		            cell.format(target_.format(std::stoull(parser().attribute("s"))));
+                cell.format(target_.format(static_cast<std::size_t>(std::stoull(parser().attribute("s")))));
             }
 
             auto has_value = false;
